requirements.txt
playwright
scrapy
python-dotenv
pandas
numpy